Move core config class to core_config.py
